Add Heroku deployment capability and use env port
Added a `heroku-deploy` script to allow Heroku deployment of the `ecommerce-api` and updated `ecommerce-api-endpoint`'s `build.gradle.kts` to enable the 'stage' task. Also, updated `CommonHelper` to better handle resource reading so that it pulls resources from the classpath correctly when deployed on a platform like Heroku.

The server port is changed to use a system environment variable instead of hard-coded port number, which is necessary for Heroku deployment because it assigns port numbers dynamically.

Renamed the test method in the `EcommerceApi.kt` to provide more targeted context on the test's purpose.