Updated main class path and Heroku deploy settings
Changed the path of the main class in the build.gradle.kts file to reflect the correct address of the EcommerceApi class. On the Heroku side, the creation of the ecommerce-api-search app was updated to use Heroku's gradle buildpack instead of the default one. This was necessary to ensure that the application is built and deployed correctly on Heroku.