Added source parameter to searchProduct API
In this commit, a 'source' query parameter was added to the `searchProduct` function in the Ecommerce API. This allows clients to specify from which ecommerce source they want to search products. If source is not provided, the function will use a default combining search. The `EcommerceSource` model was also imported to handle the possible ecommerce source values.