"Improve scraping privacy and efficiency
Added the 'puppeteer-extra-plugin-anonymize-ua' as a dependency to increase scraping privacy by anonymizing user-agent of the scraper. Simplified launching of puppeteer and playwright browsers, removing unnecessary properties and added setting of default timeout. Unnecessary comments and console.logs were also cleaned up to streamline the process. These changes are intended to increase the efficiency and privacy of our scraping engine."